Waaait that was not thinking clearly. As we go with a consistent accumulator adding values add them to a hashmap of pending values to be merged based on the rdd id & partition id - then driver side when merging the accumulators check each rdd & partition id individually against the rdd & bitset of processed partitions for that rdd adding to a single accumulator. (This way we don't need to keep all of the values around forever but if we say have first() and then foreach() we can add the value for partition one then when the second task happens and we can skip just the adding partion one. Up next: start adding some simple tests for this to see if it works

+++ core/src/main/scala/org/apache/spark/TaskContext.scala
@@ -196,10 +196,10 @@ abstract class TaskContext extends Serializable {
   /**
    * Set IDs of RDD and Split being computed. This is used by consistent accumulators.
    */
-  private[spark] def setComputeRDDSplit(rddId: Long, splitId: Long): Unit
+  private[spark] def setComputeRDDSplit(rddId: Int, splitId: Int): Unit
 
   /**
    * Return the IDs of RDD and Split being computed. This is used by consistent accumulators.
    */
-  private[spark] def getComputeRDDSplit(): (Long, Long)
+  private[spark] def getComputeRDDSplit(): (Int, Int)
 }
